Multimodel-query-RAG-system/requirements.txt
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
transformers
torch
pypdf
Pillow
accelerate
bitsandbytes
faiss-cpu
tqdm